fixed a bug with the scaling variable

+++ src/world_system.hpp
@@ -87,6 +87,6 @@ private:
 	void setupWindowScaling();
 
 	// window scaling variables
-	int gameHeight;
-	int doorWidth;
+	float gameHeight;
+	float doorWidth;
 };
